// assets/Scripts/Common/Enum.ts
export enum SceneEnum {
  H1 = 'H1',
  H2 = 'H2',
  H2A = 'H2A',
  H3 = 'H3',
  H4 = 'H4',
  Menu = 'Menu',
}

export enum ItemTypeEnum {
  Key = 'Key',
  Ticket = 'Ticket',
}

export enum ItemStateEnum {
  Scene =  'Scene',
  Package = 'Package',
  Used = 'Used',
}

export enum TriggerItemTypeEnum {
  MailBox = 'MailBox',
  Door = 'Door',
  Grandma = 'Grandma',
}

export enum TriggerItemStateEnum {
  Pending = 'Pending',
  Triggered = 'Triggered',
}

export enum EventEnum {
  Render = 'Render',
}

export enum PrefabEnum {
  Key = 'Key',
  Ticket = 'Ticket',
}